+++ src/test/java/scs/comp5903/cucumber/integration/CmsSimpleRegisterCourseStepDefs.java
@@ -1,12 +1,12 @@
-package scs.comp5903.cucumber.sample;
+package scs.comp5903.cucumber.integration;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.slf4j.Logger;
+import scs.comp5903.cucumber.integration.cmsutil.CmsPageUtils;
+import scs.comp5903.cucumber.integration.cmsutil.Constants;
+import scs.comp5903.cucumber.integration.cmsutil.SeleniumFactory;
 import scs.comp5903.cucumber.model.annotation.step.JStep;
-import scs.comp5903.cucumber.sample.cmsutil.CmsPageUtils;
-import scs.comp5903.cucumber.sample.cmsutil.Constants;
-import scs.comp5903.cucumber.sample.cmsutil.SeleniumFactory;
 
 import java.util.concurrent.CountDownLatch;
